+++ frontend/src/components/Globals/AppbarGlobal/index.tsx
@@ -1,7 +1,7 @@
 import { lazy, Suspense, useCallback, useEffect, useReducer, useState } from 'react';
 import { Container, Navbar } from 'react-bootstrap';
-import Logotype from 'components/Util/Logotype';
-import GenericSkeleton from 'components/Util/GenericSkeleton';
+import Logotype from 'components/Utils/Logotype';
+import GenericSkeleton from 'components/Utils/GenericSkeleton';
 import { useSolidityContractProvider } from 'providers/useSolidityContractProvider';
 import { TypeMetaMaskStorageData } from 'hooks/useStorageDB';
