+++ frontend/src/components/Utils/Logotype/index.test.tsx
