+++ frontend/src/components/Utils/Logotype/index.tsx
